feat(desktop): drop the rail-footer build stamp
The rail footer showed the running build next to Settings — `v0.2.0` on a
release, `v0.2.0 · a1b2c3d` on a dev build — so a developer could tell at a
glance whether the app in front of them was the installed release or the
tree they had just built (#3430; the motivating incident was mistaking one
for the other and concluding a merged change had not landed).

Build identity is not a high-frequency question, and a permanent front-stage
label is the wrong price for it — on a release build the stamp is pure
clutter, and the dev-build confusion it guarded against is a workflow
problem the About page already answers. That page has carried the same
information since PR-BUILD-HYGIENE-0: version badge, dev-build badge with
the short commit, and update state. It stays the one home for it.

The whole stamp path goes with the element: the `SidebarBuildStamp` prop on
`SessionSidebarFooter`/`SessionListPanel`, the renderer's `useBuildStamp`
IPC read, the shell-copy entry, the rail CSS, and the stamp tests. The
footer keeps its update-reminder coverage — renamed to
`sidebar-footer.test.tsx` — since that control is unchanged. The
main-process `build-info → app:info` authority stays: About still reads it,
and the update service uses `buildMode` to short-circuit dev builds.

Generated-by: Maka
