+++ packages/ui/src/session-list-panel.tsx
@@ -34,7 +34,6 @@ import {
 import {
   SessionSidebarFooter,
   SessionSidebarNav,
-  type SidebarBuildStamp,
   type SidebarUpdateReminder,
 } from './session-sidebar-nav.js';
 import { useUiLocale } from './locale-context.js';
@@ -73,7 +72,6 @@ export function SessionListPanel(props: {
   moduleMemory?: NavModuleMemory;
   onSelect(selection: NavSelection): void;
   onOpenSettings(): void;
-  buildStamp?: SidebarBuildStamp;
   updateReminder?: SidebarUpdateReminder;
   onOpenUpdate?(): void;
   onNew(): void;
@@ -182,7 +180,6 @@ export function SessionListPanel(props: {
         }
         footer={
           <SessionSidebarFooter
-            buildStamp={props.buildStamp}
             updateReminder={props.updateReminder}
             onOpenSettings={props.onOpenSettings}
             onOpenUpdate={props.onOpenUpdate}
